requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==2.30.0
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
